% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_plumber.R
\name{kill_plumber}
\alias{kill_plumber}
\title{Kill plumber}
\usage{
kill_plumber(trml_id = NULL)
}
\arguments{
\item{trml_id}{Terminal ID to kill}
}
\description{
By default, kills any terminal sessions previously started by \link{start_plumber}. To kill another terminal session, check current open terminals with \link[rstudioapi]{terminalList}.
}
