% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_plumber.R
\name{start_plumber}
\alias{start_plumber}
\title{Start plumber}
\usage{
start_plumber(path, port)
}
\arguments{
\item{path}{Path to the plumber file}

\item{port}{Port}
}
\description{
Runs a plumber API in a new terminal for interactive testing.
}
